+++ .gitignore
@@ -1,2 +1,3 @@
+radiolabdocker/__pycache__/*
 /docker-compose.yml
 /build
